# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 3.5.0
  description: >-
    (Sovereign) Agent Governed Experience — REST API for memory submission,
    validation, and query.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
...
  schemas:
...
    RegisteredAgent:
      type: object
      properties:
        agent_id:
          type: string
          description: Hex-encoded Ed25519 public key.
        name:
          type: string
        role:
          type: string
        boot_bio:
          type: string
        provider:
          type: string
        status:
          type: string
        clearance:
          type: integer
        on_chain_height:
          type: integer
          format: int64
          description: Block height where registered on-chain.
        visible_agents:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.

  # -----------------------------------------------------------------------
  # Reusable responses
  # -----------------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/agent/{id}:
    get:
      operationId: getRegisteredAgent
      summary: Get a registered agent by ID.
      tags:
        - Agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Agent's hex-encoded Ed25519 public key.
      responses:
        "200":
          description: Agent record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredAgent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...